filter and charts

+++ app/controllers/reports_controller.rb
@@ -1,5 +1,6 @@
 class ReportsController < ApplicationController
   before_action :init
+  before_action :set_session, only: %w[all_operation cards]
   protect_from_forgery except: :show
 
   def cards
@@ -9,20 +10,21 @@ class ReportsController < ApplicationController
   end
 
   def all_operation
-    session[:form_date] = params.require(:form_date).permit(:type, :start, :stop) if params.keys.include? 'form_date'
-    session[:sql] = sql_build unless (params.keys & %w[simple product value]).empty?
-    if params.keys.include? 'clear'
-      session.delete(:sql)
-      session.delete(:form_date)
-    end
-    @form_date = FormDate.new(session[:form_date])
-    @sql = session[:sql]
     @lifts = @lifts.where('amount < 0').where(sql_date).where(@sql)
     @chart_data = chart_data @lifts
     @lifts = @lifts.order(date_of_commissioned: :desc)
   end
 
   private
+
+  def set_session
+    session[:form_date] = params.require(:form_date).permit(:type, :start, :stop) if params.keys.include? 'form_date'
+    session[:sql] = sql_build unless (params.keys & %w[simple product value]).empty?
+    session.delete(:sql) if params.keys.include? 'clear_sql'
+    session.delete(:form_date) if params.keys.include? 'clear_date'
+    @form_date = FormDate.new(session[:form_date])
+    @sql = session[:sql]
+  end
 
   def sql_date
     case @form_date.type
@@ -51,7 +53,11 @@ class ReportsController < ApplicationController
   end
 
   def get_line_data(lifts)
-    lifts.group(:date_of_commissioned, :amount).pluck(:date_of_commissioned, :amount).map do |lift|
+    # lifts = lifts.group_by {|t| t.date_of_commissioned.beginning_of_month}
+
+    lifts.group('EXTRACT(month from date_of_commissioned)')
+        .order('extract_month_from_date_of_commissioned asc')
+        .sum(:amount).map do |lift|
       lift[1] = lift[1].abs
       lift
     end
